backend/requirements.txt
alembic~=1.13.1
fastapi[standard]~=0.115.0
pydantic[email]~=2.7.1
pydantic-extra-types~=2.8.0
requests~=2.32.1
psycopg2-binary~=2.9.9
sentry-sdk[fastapi,sqlalchemy,starlette]~=2.8.0
sqlalchemy~=2.0.30
pyjwt[crypto]~=2.8.0
pyotp~=2.9.0
passlib~=1.7.4
bcrypt~=4.1.3
pandas~=2.2.2
fastapi-pagination~=0.12.31
fastapi-filter[sqlalchemy]~=2.0.0
sendgrid~=6.11.0
python-dotenv~=1.0.1
tunsberg~=0.1.4
phonenumbers~=8.13.48
postmarker~=1.0.0
